axios header changed to accept/send cookies for token, added confirm for deleting taskboards

+++ tackyboard-frontend/src/NavBar/index.js
@@ -1,3 +1,4 @@
+import axios from 'axios';
 import React from 'react';
 import { useHistory } from 'react-router-dom';
 import "./NavBar.css";
@@ -6,11 +7,15 @@ function NavBar({ isLoggedIn, setLoggedIn }) {
   const history = useHistory();
 
   // handleLogout is placed here because `useHistory` can only be used within `BrowserRouter`
-  const handleLogout = () => {
-    setLoggedIn(false);
-    localStorage.clear();
-    // dispatch set redux state to INITIAL
-    history.push("/");
+  const handleLogout = async () => {
+    
+    const resp = await axios.get("http://localhost:5000/logout", {withCredentials: true});
+    if (resp.status === 200) {
+      setLoggedIn(false);
+      localStorage.clear();
+      // dispatch set redux state to INITIAL
+      history.push("/");
+    }
   }
 
   return (
